[frontend] - chore(customdashboards) refacto with common components  #925
