+++ apps/portal-front/src/components/service/custom-dashboards/[serviceInstanceId]/custom-dashboard-form.tsx
@@ -1,366 +1,558 @@
 import { getLabels } from '@/components/admin/label/label.utils';
 import { PortalContext } from '@/components/me/app-portal-context';
+import { ServiceCapabilityName } from '@/components/service/[slug]/capabilities/capability.helper';
+import { useServiceContext } from '@/components/service/components/service-context';
+import { ServiceDelete } from '@/components/service/components/service-delete';
+import FileInputWithPrevent from '@/components/ui/file-input-with-prevent';
 import MarkdownInput from '@/components/ui/MarkdownInput';
 import SelectUsersFormField from '@/components/ui/select-users';
 import { useDialogContext } from '@/components/ui/sheet-with-preventing-dialog';
-import { zodResolver } from '@hookform/resolvers/zod';
+import useServiceCapability from '@/hooks/useServiceCapability';
+import { fileToBase64 } from '@/lib/utils';
 import {
+  docIsExistingFile,
+  ExistingFile,
+  fileListCheck,
+  NewFile,
+} from '@/utils/documents';
+import { SubscribableResource } from '@/utils/shareable-resources/shareable-resources.types';
+import { AddIcon, DeleteIcon, ReplayIcon } from 'filigran-icon';
+import {
+  AutoForm,
   Button,
-  Checkbox,
   FileInput,
-  Form,
   FormControl,
-  FormField,
   FormItem,
   FormLabel,
   FormMessage,
-  Input,
   MultiSelectFormField,
+  Select,
+  SelectContent,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
   SheetFooter,
-  Textarea,
+  Tooltip,
+  TooltipContent,
+  TooltipProvider,
+  TooltipTrigger,
 } from 'filigran-ui';
 import { useTranslations } from 'next-intl';
-import { useContext, useEffect } from 'react';
-import { useForm } from 'react-hook-form';
+import { ChangeEvent, useContext, useMemo, useRef, useState } from 'react';
 import slugify from 'slugify';
 import { z } from 'zod';
 
-const fileListCheck = (file: FileList | undefined) => file && file.length > 0;
-
-export const newCustomDashboardSchema = z.object({
+const customDashboardSchema = z.object({
   name: z.string().min(1, 'Required'),
+  slug: z.string().min(1, 'Required'),
+  uploader_id: z.string().optional(),
   short_description: z.string().max(255).min(1, 'Required'),
   description: z.string().min(1, 'Required'),
   product_version: z.string().regex(/^\d+\.\d+\.\d+$/, {
     message: 'Product version must be X.Y.Z',
   }),
-  documentId: z.string().optional(),
+  uploader_organization_id: z.string().min(1, 'Required'),
+  labels: z.array(z.string()).optional(),
+  active: z.boolean().optional(),
   document: z.custom<FileList>(fileListCheck),
   images: z.custom<FileList>(fileListCheck),
-  active: z.boolean(),
-  labels: z.array(z.string()).optional(),
-  slug: z.string().min(1, 'Required'),
-  uploader_id: z.string().optional(),
 });
 
-export type CustomDashboardFormValues = z.infer<
-  typeof newCustomDashboardSchema
->;
+export type CustomDashboardFormValues = z.infer<typeof customDashboardSchema>;
 
 interface CustomDashboardFormProps {
-  handleSubmit: (
-    values: z.infer<typeof newCustomDashboardSchema>,
-    callback: () => void
-  ) => void;
+  document: SubscribableResource | undefined;
+  handleSubmit: (values: CustomDashboardFormValues) => void;
+  onDelete: () => void;
+  userCanDelete: boolean;
 }
 
 export const CustomDashboardForm = ({
+  document,
   handleSubmit,
+  userCanDelete,
+  onDelete,
 }: CustomDashboardFormProps) => {
+  const customDashboard = document;
   const t = useTranslations();
-  const { me } = useContext(PortalContext);
+  const { translationKey, serviceInstance } = useServiceContext();
   const { handleCloseSheet, setIsDirty } = useDialogContext();
 
-  const form = useForm<CustomDashboardFormValues>({
-    resolver: zodResolver(newCustomDashboardSchema),
-    criteriaMode: 'all',
-    defaultValues: {
-      name: '',
-      short_description: undefined,
-      description: '',
-      product_version: '',
-      documentId: '',
-      active: false,
-      document: undefined,
-      images: undefined,
-      labels: [],
-      slug: '',
-      uploader_id: me?.id,
-    },
-  });
+  const { me } = useContext(PortalContext);
 
-  useEffect(() => setIsDirty(form.formState.isDirty), [form.formState.isDirty]);
-  form.watch(['images', 'document']);
+  const isCreation = !customDashboard;
 
-  const onSubmit = (values: z.infer<typeof newCustomDashboardSchema>) => {
-    handleSubmit(values, () => form.reset());
-  };
+  const onSubmit = (values: CustomDashboardFormValues) => {
+    if (isCreation) {
+      handleSubmit(values);
+    } else {
+      const finalImages = images.filter(
+        (img) => !imagesToDelete.includes(img.id)
+      );
 
-  const handleNameChange = (value: string) => {
-    if (!form.formState.dirtyFields.slug) {
-      const generatedSlug = slugify(value, { lower: true, strict: true });
-      form.setValue('slug', generatedSlug, { shouldDirty: false });
+      const finalValues = {
+        ...values,
+        images: finalImages as unknown as FileList,
+      };
+      handleSubmit(finalValues);
     }
   };
 
+  const [imagesToDelete, setImagesToDelete] = useState<string[]>([]);
+
+  const [images, setImages] = useState<Array<ExistingFile | NewFile>>(
+    customDashboard?.children_documents as ExistingFile[]
+  );
+
+  const userCanUpdate = useServiceCapability(
+    ServiceCapabilityName.Upload,
+    serviceInstance
+  );
+
+  const values = useMemo(
+    () =>
+      ({
+        ...customDashboard,
+        images: customDashboard?.children_documents?.map((doc) => ({
+          ...doc,
+          name: doc.file_name,
+        })) as unknown as FileList,
+        labels: customDashboard?.labels?.map((label) => label.id),
+        uploader_id: customDashboard?.uploader?.id ?? me?.id,
+        uploader_organization_id:
+          customDashboard?.uploader_organization?.id ?? '',
+      }) as CustomDashboardFormValues,
+    [me, customDashboard]
+  );
+  const formSchema = useMemo(
+    () =>
+      customDashboard
+        ? customDashboardSchema.merge(
+            z.object({
+              document: z.custom<FileList>(fileListCheck).optional(),
+              images: z.custom<FileList>(fileListCheck).optional(),
+            })
+          )
+        : customDashboardSchema,
+    [customDashboard]
+  );
+
   return (
-    <>
-      <Form {...form}>
-        <form
-          onSubmit={form.handleSubmit(onSubmit)}
-          className="w-full space-y-xl">
-          <FormField
-            control={form.control}
-            name="name"
-            render={({ field }) => (
-              <FormItem>
-                <FormLabel>
-                  {t('Service.CustomDashboards.Form.NameLabel')}
-                </FormLabel>
-                <FormControl>
-                  <Input
-                    placeholder={t(
-                      'Service.CustomDashboards.Form.NamePlaceholder'
-                    )}
-                    {...field}
-                    onChange={(e) => {
-                      field.onChange(e);
-                      handleNameChange(e.target.value);
-                      return true;
-                    }}
-                  />
-                </FormControl>
-                <FormMessage />
-              </FormItem>
-            )}
-          />
-          <FormField
-            control={form.control}
-            name="slug"
-            render={({ field }) => (
-              <FormItem>
-                <FormLabel>
-                  {t('Service.CustomDashboards.Form.SlugLabel')}
-                </FormLabel>
-                <FormControl>
-                  <Input
-                    placeholder={t(
-                      'Service.CustomDashboards.Form.SlugPlaceholder'
-                    )}
-                    {...field}
-                  />
-                </FormControl>
-                <FormMessage />
-              </FormItem>
-            )}
-          />
-
-          <FormField
-            control={form.control}
-            name="short_description"
-            render={({ field }) => (
-              <FormItem>
-                <FormLabel>
-                  {t('Service.CustomDashboards.Form.ShortDescriptionLabel')}
-                </FormLabel>
-                <FormControl>
-                  <Textarea
-                    placeholder={t(
-                      'Service.CustomDashboards.Form.ShortDescriptionPlaceholder'
-                    )}
-                    maxLength={250}
-                    {...field}
-                  />
-                </FormControl>
-                <FormMessage />
-              </FormItem>
-            )}
-          />
-
-          <FormField
-            control={form.control}
-            name="product_version"
-            render={({ field }) => (
-              <FormItem>
-                <FormLabel>
-                  {t('Service.CustomDashboards.Form.productVersion')}
-                </FormLabel>
-                <FormControl>
-                  <Input
-                    placeholder={t(
-                      'Service.CustomDashboards.Form.productVersionPlaceholder'
-                    )}
-                    {...field}
-                  />
-                </FormControl>
-                <FormMessage />
-              </FormItem>
-            )}
-          />
-
-          <FormField
-            control={form.control}
-            name="description"
-            render={({ field }) => (
-              <FormItem>
-                <FormLabel>
-                  {t('Service.CustomDashboards.Form.DescriptionLabel')}
-                </FormLabel>
-                <FormControl>
-                  <MarkdownInput
-                    value={field.value}
-                    onChange={field.onChange}
-                    placeholder={
-                      'Service.CustomDashboards.Form.DescriptionPlaceholder'
+    <TooltipProvider delayDuration={1}>
+      <AutoForm
+        onSubmit={(values, _methods) => {
+          onSubmit(values as CustomDashboardFormValues);
+        }}
+        onValuesChange={(values, form) => {
+          if (values.name) {
+            const generatedSlug = slugify(values.name, {
+              lower: true,
+              strict: true,
+            });
+            const currentSlug = form.getValues('slug');
+            if (currentSlug !== generatedSlug) {
+              form.setValue('slug', generatedSlug, { shouldDirty: false });
+            }
+          }
+          if (!isCreation && values.images) {
+            form.setValue('images', images as unknown as FileList);
+          }
+          if (isCreation) {
+            form.setValue(
+              'uploader_organization_id',
+              me!.selected_organization_id
+            );
+          }
+        }}
+        values={values}
+        formSchema={formSchema}
+        fieldConfig={{
+          ...(userCanUpdate
+            ? {
+                description: {
+                  fieldType: ({ field }) => (
+                    <FormItem>
+                      <FormLabel>
+                        {t('Service.CustomDashboards.Form.DescriptionLabel')}
+                      </FormLabel>
+                      <FormControl>
+                        <MarkdownInput
+                          value={field.value}
+                          onChange={field.onChange}
+                          placeholder={
+                            'Service.CustomDashboards.Form.DescriptionPlaceholder'
+                          }
+                        />
+                      </FormControl>
+                      <FormMessage />
+                    </FormItem>
+                  ),
+                },
+                labels: {
+                  fieldType: ({ field }) => (
+                    <FormItem>
+                      <FormLabel>
+                        {t('Service.CustomDashboards.Form.LabelsLabel')}
+                      </FormLabel>
+                      <FormControl>
+                        <MultiSelectFormField
+                          noResultString={t('Utils.NotFound')}
+                          options={getLabels()}
+                          keyValue="id"
+                          keyLabel="name"
+                          defaultValue={field.value}
+                          value={field.value}
+                          onValueChange={field.onChange}
+                          placeholder={t(
+                            'Service.CustomDashboards.Form.LabelsPlaceholder'
+                          )}
+                          variant="inverted"
+                        />
+                      </FormControl>
+                    </FormItem>
+                  ),
+                },
+                uploader_id: {
+                  fieldType: ({ field }) => (
+                    <FormItem>
+                      <FormLabel>
+                        {t('Service.CustomDashboards.Form.Author')}
+                      </FormLabel>
+                      <FormControl>
+                        <SelectUsersFormField
+                          defaultValue={
+                            customDashboard?.uploader?.email ?? me!.email
+                          }
+                          value={field.value}
+                          onValueChange={field.onChange}
+                        />
+                      </FormControl>
+                    </FormItem>
+                  ),
+                },
+                uploader_organization_id: {
+                  fieldType: ({ field }) => (
+                    <FormItem hidden={isCreation}>
+                      <FormLabel>
+                        {t('OrganizationInServiceAction.Organization')}
+                      </FormLabel>
+                      <Select
+                        onValueChange={field.onChange}
+                        defaultValue={field.value}>
+                        <FormControl>
+                          <SelectTrigger>
+                            <SelectValue
+                              placeholder={t(
+                                'OrganizationInServiceAction.SelectOrganization'
+                              )}
+                            />
+                          </SelectTrigger>
+                        </FormControl>
+                        <SelectContent>
+                          {me?.organizations.map((node) => {
+                            return (
+                              <SelectItem
+                                key={node?.id}
+                                value={node?.id}>
+                                {node?.name}
+                              </SelectItem>
+                            );
+                          })}
+                        </SelectContent>
+                      </Select>
+                      <FormMessage />
+                    </FormItem>
+                  ),
+                },
+                document: isCreation
+                  ? {
+                      label: t('Service.CustomDashboards.Form.SelectJSONFile'),
+                      fieldType: 'file',
+                      inputProps: {
+                        allowedTypes: 'application/json',
+                        multiple: 'multiple',
+                      },
                     }
-                  />
-                </FormControl>
-                <FormMessage />
-              </FormItem>
-            )}
-          />
-
-          <FormField
-            control={form.control}
-            name="labels"
-            render={({ field }) => (
-              <FormItem>
-                <FormLabel>
-                  {t('Service.CustomDashboards.Form.LabelsLabel')}
-                </FormLabel>
-                <FormControl>
-                  <MultiSelectFormField
-                    noResultString={t('Utils.NotFound')}
-                    options={getLabels().map(({ name, id }) => ({
-                      label: name,
-                      value: id,
-                    }))}
-                    defaultValue={field.value}
-                    onValueChange={field.onChange}
-                    placeholder={t(
-                      'Service.CustomDashboards.Form.LabelsPlaceholder'
-                    )}
-                    variant="inverted"
-                  />
-                </FormControl>
-
-                <FormMessage />
-              </FormItem>
-            )}
-          />
-
-          <FormField
-            control={form.control}
-            name="active"
-            render={({ field }) => (
-              <FormItem>
-                <FormLabel>
-                  {t('Service.CustomDashboards.Form.PublishedLabel')}
-                </FormLabel>
-                <div className="flex items-center gap-2">
-                  <FormControl>
-                    <Checkbox
-                      {...field}
-                      checked={field.value}
-                      value="on"
-                      onCheckedChange={() =>
-                        form.setValue('active', !field.value)
-                      }
-                    />
-                  </FormControl>
-                  <FormLabel className="font-normal cursor-pointer">
-                    {t('Service.CustomDashboards.Form.PublishedPlaceholder')}
-                  </FormLabel>
-                </div>
-                <FormMessage />
-              </FormItem>
-            )}
-          />
-
-          <FormField
-            control={form.control}
-            name="uploader_id"
-            render={({ field }) => (
-              <FormItem>
-                <FormLabel>
-                  {t('Service.CustomDashboards.Form.Author')}
-                </FormLabel>
-                <FormControl>
-                  <SelectUsersFormField
-                    defaultValue={me!.email}
-                    value={field.value}
-                    onValueChange={field.onChange}
-                  />
-                </FormControl>
-              </FormItem>
-            )}
-          />
-
-          <FormField
-            control={form.control}
-            name="document"
-            render={({ field }) => {
-              return (
-                <FormItem>
-                  <FormLabel>
-                    {t('Service.CustomDashboards.Form.JSONFile')}
-                  </FormLabel>
-                  <FormControl>
-                    <FileInput
-                      {...field}
-                      texts={{
-                        selectFile: t(
-                          'Service.CustomDashboards.Form.SelectJSONFile'
-                        ),
-                        noFile: t('Service.CustomDashboards.Form.NoJSONFile'),
-                        dropFiles: t('Service.Vault.FileForm.DropDocuments'),
-                      }}
-                      allowedTypes={'application/json'}
-                    />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              );
-            }}
-          />
-
-          <FormField
-            control={form.control}
-            name={'images'}
-            render={({ field: { value, ref } }) => (
-              <FormItem>
-                <FormLabel className="flex items-center gap-2 h-6">
-                  {t('Service.CustomDashboards.Form.ImageLabel')}
-                </FormLabel>
-                <FormControl>
-                  <FileInput
-                    multiple
-                    name={'images'}
-                    texts={{
-                      selectFile: t(
-                        'Service.CustomDashboards.Form.SelectImage'
+                  : {
+                      fieldType: ({ field }) => (
+                        <FormItem>
+                          <FormLabel>
+                            {t(
+                              'Service.CustomDashboards.Form.ExistingJSONFile',
+                              {
+                                file_name:
+                                  field.value?.[0].name ??
+                                  customDashboard?.file_name,
+                              }
+                            )}
+                          </FormLabel>
+                          <FormControl>
+                            <div onClick={() => setIsDirty(true)}>
+                              <FileInputWithPrevent
+                                field={field}
+                                texts={{
+                                  selectFile: t(
+                                    'Service.CustomDashboards.Form.UpdateJSONFile'
+                                  ),
+                                  dialogTitle: t(
+                                    'Service.CustomDashboards.Form.UpdateJSONFile'
+                                  ),
+                                  dialogDescription: t(
+                                    'Service.CustomDashboards.Form.DescriptionUpdateJSONFile'
+                                  ),
+                                }}
+                                allowedTypes="application/json"
+                              />
+                            </div>
+                          </FormControl>
+                          <FormMessage />
+                        </FormItem>
                       ),
-                      noFile: t('Service.CustomDashboards.Form.NoImage'),
-                      dropFiles: t('Service.Vault.FileForm.DropDocuments'),
-                    }}
-                    allowedTypes={'image/jpeg, image/png'}
-                    ref={ref}
-                    value={value ? [value] : []}
-                  />
-                </FormControl>
-                <FormMessage />
-              </FormItem>
-            )}
-          />
+                    },
+                images: isCreation
+                  ? {
+                      label: t('Service.CustomDashboards.Form.ImageLabel'),
+                      fieldType: 'file',
+                      inputProps: {
+                        allowedTypes: 'image/jpeg, image/png',
+                        multiple: 'multiple',
+                        texts: {
+                          selectFile: t(
+                            'Service.CustomDashboards.Form.SelectImage'
+                          ),
+                          noFile: t('Service.CustomDashboards.Form.NoImage'),
+                          dropFiles: t('Service.Vault.FileForm.DropDocuments'),
+                        },
+                      },
+                    }
+                  : {
+                      fieldType: ({ field: { value, ref } }) => {
+                        const inputRef = useRef<HTMLInputElement | null>(null);
+                        return (
+                          <>
+                            <FormItem>
+                              <FormLabel className="flex items-center h-6">
+                                {t('Service.CustomDashboards.Form.ImageLabel')}
+                                <Button
+                                  size="icon"
+                                  variant="link"
+                                  onClick={(e) => {
+                                    e.preventDefault();
+                                    inputRef.current!.click();
+                                  }}>
+                                  <AddIcon className="size-3" />
+                                </Button>
+                              </FormLabel>
+                              <FormControl>
+                                <FileInput
+                                  multiple
+                                  hidden
+                                  name="images"
+                                  onChangeCapture={async (
+                                    e: ChangeEvent<HTMLInputElement>
+                                  ) => {
+                                    const localImages = [...images];
+                                    if (e.target?.files) {
+                                      for (const image of Array.from(
+                                        e.target.files
+                                      )) {
+                                        const extendedImage = image as NewFile;
+                                        extendedImage.preview =
+                                          await fileToBase64(image as File);
+                                        extendedImage.id = new Date()
+                                          .getTime()
+                                          .toString();
+                                        localImages.push(extendedImage);
+                                      }
+                                    }
+                                    setImages(localImages);
+                                    return false;
+                                  }}
+                                  texts={{
+                                    selectFile: t(
+                                      'Service.CustomDashboards.Form.UploadImage'
+                                    ),
+                                    noFile: t(
+                                      'Service.CustomDashboards.Form.NoImage'
+                                    ),
+                                    dropFiles: t(
+                                      'Service.Vault.FileForm.DropDocuments'
+                                    ),
+                                  }}
+                                  allowedTypes={'image/jpeg, image/png'}
+                                  ref={(e: HTMLInputElement) => {
+                                    ref(e);
+                                    inputRef.current = e;
+                                  }}
+                                  value={value ? [value] : []}
+                                />
+                              </FormControl>
+                              <FormMessage />
+                            </FormItem>
+                            {images?.length > 0 && (
+                              <div
+                                className="grid grid-cols-1 s:grid-cols-2 md:grid-cols-2 lg:grid-cols-3 2xl:grid-cols-3 gap-xl min-h-[15rem] pb-xl"
+                                data-testid="images-grid">
+                                {images.map((doc) => (
+                                  <div
+                                    key={doc!.id}
+                                    style={{
+                                      backgroundImage: docIsExistingFile(doc)
+                                        ? `url(/document/visualize/${customDashboard!.service_instance!.id}/${doc!.id})`
+                                        : `url(${doc.preview})`,
+                                      backgroundSize: 'cover',
+                                    }}
+                                    className="min-h-[15rem] border rounded relative">
+                                    <div
+                                      className={`absolute inset-0 bg-black flex flex-col items-center justify-center transition-all duration-800 ease-in ${
+                                        imagesToDelete.includes(doc!.id)
+                                          ? 'bg-opacity-90 opacity-100'
+                                          : 'bg-opacity-0 opacity-0'
+                                      }`}>
+                                      <Button
+                                        variant="outline"
+                                        size="icon"
+                                        className="absolute right-2 top-2"
+                                        type="button"
+                                        onClick={() => {
+                                          setImagesToDelete(
+                                            imagesToDelete.filter(
+                                              (id) => id !== doc!.id
+                                            )
+                                          );
+                                        }}>
+                                        <Tooltip>
+                                          <TooltipTrigger asChild>
+                                            <ReplayIcon className="size-4" />
+                                          </TooltipTrigger>
+                                          <TooltipContent className="bg-gray-50">
+                                            {t(
+                                              'Service.CustomDashboards.Form.Restore'
+                                            )}
+                                          </TooltipContent>
+                                        </Tooltip>
+                                      </Button>
+                                      <DeleteIcon
+                                        focusable={false}
+                                        className="size-6 text-gray-300"
+                                      />
+                                    </div>
 
-          <SheetFooter className="pt-2">
-            <div className="flex gap-s">
-              <Button
-                variant="outline"
-                type="button"
-                onClick={(e) => handleCloseSheet(e)}>
-                {t('Utils.Cancel')}
-              </Button>
-
-              <Button
-                disabled={!form.formState.isDirty}
-                type="submit">
-                {t('Utils.Validate')}
-              </Button>
-            </div>
-          </SheetFooter>
-        </form>
-      </Form>
-    </>
+                                    {!imagesToDelete.includes(doc!.id) && (
+                                      <div className="flex flex-row items-center bg-page-background h-12 opacity-90">
+                                        <div className="truncate overflow-hidden whitespace-nowrap text-ellipsis ml-s mr-s flex-1 min-w-0">
+                                          {(doc as ExistingFile)?.file_name ??
+                                            (doc as NewFile)?.name}
+                                        </div>
+                                        <Button
+                                          variant="outline-destructive"
+                                          size="icon"
+                                          type="button"
+                                          className="ml-auto m-s"
+                                          onClick={() => {
+                                            setImagesToDelete([
+                                              ...imagesToDelete,
+                                              doc!.id,
+                                            ]);
+                                            setIsDirty(true);
+                                          }}>
+                                          <Tooltip>
+                                            <TooltipTrigger asChild>
+                                              <DeleteIcon className="size-4" />
+                                            </TooltipTrigger>
+                                            <TooltipContent className="bg-gray-50">
+                                              {t(
+                                                'Service.CustomDashboards.Form.DeleteSentence'
+                                              )}
+                                            </TooltipContent>
+                                          </Tooltip>
+                                        </Button>
+                                      </div>
+                                    )}
+                                  </div>
+                                ))}
+                              </div>
+                            )}
+                          </>
+                        );
+                      },
+                    },
+                active: {
+                  label: t(
+                    'Service.CustomDashboards.Form.PublishedPlaceholder'
+                  ),
+                },
+                short_description: {
+                  label: t(
+                    'Service.CustomDashboards.Form.ShortDescriptionLabel'
+                  ),
+                  inputProps: {
+                    placeholder: t(
+                      'Service.CustomDashboards.Form.ShortDescriptionPlaceholder'
+                    ),
+                  },
+                },
+                slug: {
+                  label: t('Service.CustomDashboards.Form.SlugLabel'),
+                  inputProps: {
+                    placeholder: t(
+                      'Service.CustomDashboards.Form.SlugPlaceholder'
+                    ),
+                  },
+                },
+                name: {
+                  label: t('Service.CustomDashboards.Form.NameLabel'),
+                  inputProps: {
+                    placeholder: t(
+                      'Service.CustomDashboards.Form.NamePlaceholder'
+                    ),
+                  },
+                },
+                product_version: {
+                  label: t('Service.CustomDashboards.Form.productVersion'),
+                  inputProps: {
+                    placeholder: t(
+                      'Service.CustomDashboards.Form.productVersionPlaceholder'
+                    ),
+                  },
+                },
+              }
+            : {
+                // Used for user that have the Delete Capa but not the Upload Capa
+                name: { fieldType: () => <></> },
+                slug: { fieldType: () => <></> },
+                uploader_id: { fieldType: () => <></> },
+                short_description: { fieldType: () => <></> },
+                description: { fieldType: () => <></> },
+                product_version: { fieldType: () => <></> },
+                message: { fieldType: () => <></> },
+                uploader_organization_id: { fieldType: () => <></> },
+                labels: { fieldType: () => <></> },
+                active: { fieldType: () => <></> },
+                document: { fieldType: () => <></> },
+                images: { fieldType: () => <></> },
+              }),
+        }}>
+        <SheetFooter className="sm:justify-between pb-0">
+          {customDashboard && (
+            <ServiceDelete
+              userCanDelete={userCanDelete}
+              onDelete={onDelete}
+              serviceName={customDashboard.name}
+              translationKey={translationKey}
+            />
+          )}
+          <div className="ml-auto flex gap-s">
+            <Button
+              variant="outline"
+              type="button"
+              onClick={(e) => handleCloseSheet(e)}>
+              {t('Utils.Cancel')}
+            </Button>
+            <Button type="submit">{t('Utils.Validate')}</Button>
+          </div>
+        </SheetFooter>
+      </AutoForm>
+    </TooltipProvider>
   );
 };
