+++ apps/portal-front/src/components/service/custom-dashboards/[slug]/custom-dashboard-details.tsx
@@ -1,5 +1,7 @@
-import DashboardUpdate from '@/components/service/custom-dashboards/[serviceInstanceId]/custom-dashboard-update';
+import { AppServiceContext } from '@/components/service/components/service-context';
+import { ServiceManageSheet } from '@/components/service/components/service-manage-sheet';
 import DashboardCarousel from '@/components/service/custom-dashboards/[slug]/custom-dashboard-carousel-view';
+import { useCustomDashboardsContext } from '@/components/service/custom-dashboards/use-custom-dashboards-context';
 import ShareableResourceSlug from '@/components/service/document/shareable-resource-slug';
 import { APP_PATH } from '@/utils/path/constant';
 import { customDashboardQuery } from '@generated/customDashboardQuery.graphql';
@@ -47,22 +49,25 @@ const DashboardSlug: React.FunctionComponent<DashboardSlugProps> = ({
     },
   ];
 
+  const context = useCustomDashboardsContext(serviceInstance);
   return (
     documentData && (
-      <ShareableResourceSlug
-        breadcrumbValue={breadcrumbValue}
-        documentData={documentData}
-        updateActions={
-          <DashboardUpdate
-            serviceInstance={serviceInstance}
-            customDashboard={documentData!}
-          />
-        }>
-        <DashboardCarousel
-          serviceInstance={serviceInstance}
+      <AppServiceContext {...context}>
+        <ShareableResourceSlug
+          breadcrumbValue={breadcrumbValue}
           documentData={documentData}
-        />
-      </ShareableResourceSlug>
+          updateActions={
+            <ServiceManageSheet
+              document={documentData}
+              variant={'button'}
+            />
+          }>
+          <DashboardCarousel
+            serviceInstance={serviceInstance}
+            documentData={documentData}
+          />
+        </ShareableResourceSlug>
+      </AppServiceContext>
     )
   );
 };
